Add passives :panda_face:
Adds a new script to the HUD root prim called "got Passives". PCs will
reroute the FXCEvt through got Passives and the result will be a
combination of both through PassivesEvt$data. The parameters are exactly
the same. NPCs should be left untouched but need testing.

It also adds a handful of new things you can do with passives to modify
the max value of a resource.