More improvements
- Alert$freetext can now supply a custom sound UUID instead of
TRUE/FALSE/2 for sound.
- Added ability to add attachments to passives
- Removed some really old code from when GoThongs was meant to be turn
based.
- Added flags to passives, currently supports
Passives$FLAG_REM_ON_CLEANUP which removes a passive when a cell is
finished or a player hits the cleanup button.
- Added 3 new events to Potions: pickup, drop, use, which all contain
the name of the potion.
- Added second parameter StatusEvt$flags - The second parameter is the
flags before this update, allowing you to use bitwise operations to
easily get newly set and newly removed flags.
- Added support for FX set of team for monsters.
- Added config to got Interact to allow interactions when sitting
- PC status flags will now only update when changed
- Added fxlib$forceSit macro